Move `MatchableConcept` class out of `GraknMatchers`
